libp2p bootstrap info from file
